build: fix linux/android builds
Rather than use ubuntu:latest (presently, ubuntu:noble) use the oldest
supported version of Ubuntu for maximum compatibility (earliest glib
version, etc.) and for addressing build issues. The earliest supported
version available as a github runner at this time is ubuntu-20.04.

This was done in response to a build issue in the mini_chromium directory
where the header stdint.h needs to be included. This is addressed in later
versions of the submodule (eb418ec05f) and will get picked when this fork
is sync'd with upstream. This isn't a problem in earlier versions of GCC
because the header is implicitly included.

There were also permission issues updating ruby bundles in Ubuntu Noble
that this change works around.