build: fix macOS build in response to latest github action runner
With the latest github runner for macos-latest, macos-14-arm64
20250324.1158, the `SDKROOT` environment var is no longer set (it was in
the prior versions such as 20250317.1117).

Because of this, the CMake variable `CMAKE_OSX_SYSROOT` is not set which
causes issues when looking for various def files, such as exc.defs.

This commit updates the workflows file to set `CMAKE_OSX_SYSROOT` from
`xcrun` rather than relying on `SDKROOT` being set or not.